requirements.txt
pinecone
cohere
streamlit
python-dotenv